Address gimbal review feedback
Use runtime gimbal dispatch and typed violation indices. Consolidate the
gimbal coverage while avoiding redundant transfers and Newton-only tests.
