+++ newton/_src/solvers/kamino/_src/__init__.py
@@ -13,13 +13,7 @@ from .core.bodies import (
 )
 from .core.control import ControlKamino
 from .core.conversions import (
-    JOINT_UPDATE_VIOLATION_ACTUATION_PARTITION,
-    JOINT_UPDATE_VIOLATION_COUNT,
-    JOINT_UPDATE_VIOLATION_DYNAMIC_CTS,
-    JOINT_UPDATE_VIOLATION_GIMBAL_HANDEDNESS,
-    JOINT_UPDATE_VIOLATION_INVALID_TARGET_MODE,
-    JOINT_UPDATE_VIOLATION_LIMIT_FINITE,
-    JOINT_UPDATE_VIOLATION_NONORTHONORMAL_AXES,
+    JointUpdateViolation,
     compute_material_first_shape,
     convert_model_joint_actuation,
     convert_model_joint_transforms,
@@ -45,17 +39,11 @@ from .utils import logger as msg
 __all__ = [
     "JOINT_QMAX",
     "JOINT_QMIN",
-    "JOINT_UPDATE_VIOLATION_ACTUATION_PARTITION",
-    "JOINT_UPDATE_VIOLATION_COUNT",
-    "JOINT_UPDATE_VIOLATION_DYNAMIC_CTS",
-    "JOINT_UPDATE_VIOLATION_GIMBAL_HANDEDNESS",
-    "JOINT_UPDATE_VIOLATION_INVALID_TARGET_MODE",
-    "JOINT_UPDATE_VIOLATION_LIMIT_FINITE",
-    "JOINT_UPDATE_VIOLATION_NONORTHONORMAL_AXES",
     "CollisionDetector",
     "ContactsKamino",
     "ControlKamino",
     "JointActuationType",
+    "JointUpdateViolation",
     "ModelKamino",
     "SolverKaminoImpl",
     "StateKamino",
